+++ init.php
@@ -44,7 +44,7 @@ require(dirname(__FILE__) . '/lib/Refund.php');
 require(dirname(__FILE__) . '/lib/RedEnvelope.php');
 require(dirname(__FILE__) . '/lib/Event.php');
 require(dirname(__FILE__) . '/lib/Transfer.php');
-require(dirname(__FILE__) . '/lib/Identification.php');
+require(dirname(__FILE__) . '/lib/CardInfo.php');
 require(dirname(__FILE__) . '/lib/BatchRefund.php');
 require(dirname(__FILE__) . '/lib/BatchTransfer.php');
 require(dirname(__FILE__) . '/lib/Customs.php');